+++ Sources/Moya/AnyEncodable.swift
@@ -1,0 +1,14 @@
+import Foundation
+
+struct AnyEncodable: Encodable {
+
+    private let encodable: Encodable
+
+    public init(_ encodable: Encodable) {
+        self.encodable = encodable
+    }
+
+    func encode(to encoder: Encoder) throws {
+        try encodable.encode(to: encoder)
+    }
+}