+++ Sources/Moya/MoyaError.swift
@@ -11,6 +11,12 @@ public enum MoyaError: Swift.Error {
     /// Indicates a response failed to map to a String.
     case stringMapping(Response)
 
+    /// Indicates a response failed to map to a Decodable object.
+    case objectMapping(Swift.Error, Response)
+
+    /// Indicates that Encodable couldn't be encoded into Data
+    case encodableMapping(Swift.Error)
+
     /// Indicates a response failed with an invalid HTTP status code.
     case statusCode(Response)
 
@@ -19,6 +25,9 @@ public enum MoyaError: Swift.Error {
 
     /// Indicates that an `Endpoint` failed to map to a `URLRequest`.
     case requestMapping(String)
+
+    /// Indicates that an `Endpoint` failed to encode the parameters for the `URLRequest`.
+    case parameterEncoding(Swift.Error)
 }
 
 public extension MoyaError {
@@ -28,9 +37,12 @@ public extension MoyaError {
         case .imageMapping(let response): return response
         case .jsonMapping(let response): return response
         case .stringMapping(let response): return response
+        case .objectMapping(_, let response): return response
         case .statusCode(let response): return response
         case .underlying(_, let response): return response
+        case .encodableMapping: return nil
         case .requestMapping: return nil
+        case .parameterEncoding: return nil
         }
     }
 }
@@ -46,10 +58,16 @@ extension MoyaError: LocalizedError {
             return "Failed to map data to JSON."
         case .stringMapping:
             return "Failed to map data to a String."
+        case .objectMapping:
+            return "Failed to map data to a Decodable object."
+        case .encodableMapping:
+            return "Failed to encode Encodable object into data."
         case .statusCode:
             return "Status code didn't fall within the given range."
         case .requestMapping:
             return "Failed to map Endpoint to a URLRequest."
+        case .parameterEncoding(let error):
+            return "Failed to encode parameters for URLRequest. \(error.localizedDescription)"
         case .underlying(let error, _):
             return error.localizedDescription
         }